# Pool writeWork packets in parallel WriteTo